Autoboxing
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@24470 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/client/async/SplitFileFetcherSubSegment.java
@@ -402,7 +402,7 @@ public class SplitFileFetcherSubSegment extends SendableGet implements SupportsB
 				container.store(blockNums);
 		}
 		if(!block.isMetadata()) {
-			onSuccess(data, fromStore, (Integer)token, ((Integer)token).intValue(), block, container, context);
+			onSuccess(data, fromStore, (Integer)token, (Integer)token, block, container, context);
 		} else {
 			onFailure(new FetchException(FetchException.INVALID_METADATA, "Metadata where expected data"), token, container, context);
 			data.free();
